level editor refactors and fixes
