+++ src/editor/RoomEditingArea/roomEditingAreaMarginPx.tsx
@@ -1,0 +1,1 @@
+export const roomEditingAreaMarginPx = 40;
