+++ src/editor/RoomEditingArea/useAddRoomRendererOutputToApplicationStage.tsx
@@ -1,17 +1,19 @@
 import { Container } from "pixi.js";
 import { useState, useEffect } from "react";
-import { useAppSelector } from "../../store/hooks";
-import { selectGameEngineScreenSize } from "../../store/slices/upscale/upscaleSlice";
 import type { EditorRoomRenderer } from "../editorTypes";
 import { useProvidedPixiApplication } from "./PixiApplicationProvider";
+import { useEditorRoomRenderDimensions } from "../slice/levelEditorSelectors";
+import { roomEditingAreaMarginPx } from "./roomEditingAreaMarginPx";
 
 export const useAddRoomRendererOutputToApplicationStage = (
   roomRenderer: EditorRoomRenderer,
 ) => {
   const application = useProvidedPixiApplication();
 
-  const [positioner] = useState(() => new Container());
-  const gameEngineScreenSize = useAppSelector(selectGameEngineScreenSize);
+  const [positioner] = useState(
+    () => new Container({ label: "editorPositioner" }),
+  );
+  const roomRenderSize = useEditorRoomRenderDimensions();
 
   useEffect(() => {
     if (roomRenderer.destroyed) {
@@ -34,7 +36,7 @@ export const useAddRoomRendererOutputToApplicationStage = (
 
   // keep the room renderer's output centred on the canvas:
   useEffect(() => {
-    positioner.x = gameEngineScreenSize.x / 2;
-    positioner.y = gameEngineScreenSize.y - 16;
-  }, [positioner, gameEngineScreenSize]);
+    positioner.x = -roomRenderSize.l + roomEditingAreaMarginPx;
+    positioner.y = -roomRenderSize.t + roomEditingAreaMarginPx;
+  }, [positioner, roomRenderSize]);
 };
